move svec aggregate to single file.
