+++ src/sql/bootstrap.sql
@@ -10,7 +10,7 @@ CREATE TYPE veci8;
 
 CREATE TYPE vector_index_stat;
 
-CREATE TYPE _vectors_vecf32_aggregate_avg_stype;
-CREATE TYPE svector_accumulate_state;
+CREATE TYPE _vectors_vecf32_aggregate_stype;
+CREATE TYPE _vectors_svecf32_aggregate_stype;
 
 -- bootstrap end
